Remove analytics page, merge functionality into root page
