+++ src/components/Header.tsx
@@ -1,4 +1,4 @@
-import { Ticket, Bell, Settings, Moon, Sun, BarChart3, Home } from 'lucide-react'
+import { Ticket, Bell, Settings, Moon, Sun, BarChart3 } from 'lucide-react'
 import { useDarkMode } from '../hooks/useDarkMode'
 import { useTranslation } from 'react-i18next'
 import Link from 'next/link'
@@ -29,24 +29,12 @@ export function Header({ onNotificationClick, onSettingsClick, ...props }: Heade
                 <p className="text-sm theme-text-secondary font-medium">{t('common.tagline')}</p>
               </div>
             </Link>
-            
             {/* Navigation Links */}
             <nav className="hidden md:flex items-center gap-1">
               <Link
                 href="/"
                 className={`flex items-center gap-2 px-3 py-2 rounded-lg text-sm font-medium transition-all duration-200 ${
-                  pathname === '/' 
-                    ? 'bg-blue-100 text-blue-700 dark:bg-blue-900/30 dark:text-blue-400'
-                    : 'theme-text-secondary hover:theme-text-primary hover:bg-gray-100 dark:hover:bg-gray-800'
-                }`}
-              >
-                <Home size={16} />
-                {t('navigation.home')}
-              </Link>
-              <Link
-                href="/analytics"
-                className={`flex items-center gap-2 px-3 py-2 rounded-lg text-sm font-medium transition-all duration-200 ${
-                  pathname === '/analytics'
+                  pathname === '/'
                     ? 'bg-blue-100 text-blue-700 dark:bg-blue-900/30 dark:text-blue-400'
                     : 'theme-text-secondary hover:theme-text-primary hover:bg-gray-100 dark:hover:bg-gray-800'
                 }`}
@@ -56,25 +44,14 @@ export function Header({ onNotificationClick, onSettingsClick, ...props }: Heade
               </Link>
             </nav>
           </div>
-          
+
           <div className="flex items-center gap-1 sm:gap-2" {...props}>
             {/* Mobile Navigation */}
             <div className="md:hidden flex items-center gap-0.5 sm:gap-1">
               <Link
                 href="/"
                 className={`p-1.5 sm:p-2.5 rounded-lg transition-all duration-200 ${
-                  pathname === '/' 
-                    ? 'bg-blue-100 text-blue-700 dark:bg-blue-900/30 dark:text-blue-400'
-                    : 'text-gray-500 hover:text-gray-700 dark:text-gray-400 dark:hover:text-gray-200 hover:bg-gray-100 dark:hover:bg-gray-800'
-                }`}
-                aria-label={t('navigation.home')}
-              >
-                <Home size={16} className="sm:w-5 sm:h-5" />
-              </Link>
-              <Link
-                href="/analytics"
-                className={`p-1.5 sm:p-2.5 rounded-lg transition-all duration-200 ${
-                  pathname === '/analytics'
+                  pathname === '/'
                     ? 'bg-blue-100 text-blue-700 dark:bg-blue-900/30 dark:text-blue-400'
                     : 'text-gray-500 hover:text-gray-700 dark:text-gray-400 dark:hover:text-gray-200 hover:bg-gray-100 dark:hover:bg-gray-800'
                 }`}
@@ -83,7 +60,7 @@ export function Header({ onNotificationClick, onSettingsClick, ...props }: Heade
                 <BarChart3 size={16} className="sm:w-5 sm:h-5" />
               </Link>
             </div>
-            
+
             <button
               onClick={() => setThemeMode(isDark ? 'light' : 'dark')}
               className="p-1.5 sm:p-2.5 text-gray-500 hover:text-gray-700 dark:text-gray-400 dark:hover:text-gray-200 hover:bg-gray-100 dark:hover:bg-gray-800 rounded-lg transition-all duration-200"
@@ -100,7 +77,7 @@ export function Header({ onNotificationClick, onSettingsClick, ...props }: Heade
             >
               <Bell size={16} className="sm:w-5 sm:h-5" />
             </button>
-            <button 
+            <button
               onClick={onSettingsClick}
               className="p-1.5 sm:p-2.5 text-gray-500 hover:text-gray-700 dark:text-gray-400 dark:hover:text-gray-200 hover:bg-gray-100 dark:hover:bg-gray-800 rounded-lg transition-all duration-200"
               aria-label={t('header.settings')}
